+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.21.4
 
+* moar tests for CPManager